% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSAEZIP.R
\docType{data}
\name{dataSAEZIP}
\alias{dataSAEZIP}
\title{Sample Data for Small Area Estimation with Zero-Inflated Poisson model}
\format{
A data frame with 300 rows and 3 variables:
\describe{
  \item{y}{Direct Estimation of y}
  \item{x1}{Auxiliary variable of x1}
  \item{vardir}{Sampling Variance of y}
}
}
\usage{
dataSAEZIP
}
\description{
A Dataset which is generate with Zero-Inflated Poisson method for Small Area Estimation purpose

This data is generated based on Zero-Inflated Poisson with EBLUP based model
}
\keyword{datasets}
